Implement deletion, tuples
